Move some code to a utility class
